refactor slt schemas into module; more coverage for as_node tests;
