+++ tests/test_schema_model_source_logic_tree.py
@@ -3,12 +3,13 @@ from graphene.test import Client
 
 from nshm_model_graphql_api import schema
 
-from graphql_relay import to_global_id
+# from graphql_relay import to_global_id
 
 
 @pytest.fixture(scope="module")
 def client():
     return Client(schema.schema_root)
+
 
 @pytest.mark.parametrize(
     "model_version",
